code cleaned, props not working
