+++ modules/ListCompHooks.js
@@ -20,37 +20,44 @@ const EmptyListComponent =()=>(
   );
 
 export default ListCompHooks = (props) =>{
-    const [data, setData] = useState(null);
-    const [dataRequest, setDataRequest] = useState(false);
-    const [isLoading, setIsLoading] = useState(false);
+    const [isLoading, setIsLoading] = useState(props.isLoading);
     const [showSearchBar, setShowSearchBar] = useState(false);
     const [filterText, setFilterText] = useState('');
-    const [bottomMenu, setBottomMenu] = useState(null);
-    const [filterData, setFilterData] = useState(null);
+    const [filterData, setFilterData] = useState(props.data);   
 
-    const fetchApiData=()=>{
-        setIsLoading(true);
-        fetch(props.url)
-            .then(response => response.json())
-                .then(responseJson => {
-                    setIsLoading(false);
-                    setData(responseJson);
-                    setFilterData(responseJson);
-        })
-        .catch((error) => console.warn(error));
+    //sorting
+
+    const sortDataById=()=>{
+        // console.warn(props.data[0].author);
+        const newData = [...props.data];
+        newData.sort((a,b) => a.id - b.id);
+        setFilterData(newData);
     }
 
-    useEffect(()=>{
-        fetchApiData();
-        // console.warn("fetchuje");
-    }, [dataRequest]);
+    const sortDataByAuthor=()=>{
+        const newData = [...props.data];
+        newData.sort((a,b) => {
+            if(a.author > b.author) return 1;
+            else if(a.author < b.author) return -1;
+            else return 0;
+        });
+        setFilterData(newData);
+    }    
 
-    useEffect(()=>{
-        // console.warn("showBar");
-        setBottomMenu(showSearchBar
+    //filtering
+    const filterByAuthor=(text)=>{
+        const newData = [...props.data];
+        const newFilteredData = newData.filter(item =>{
+            const textData = text.toUpperCase();
+            const itemData = item.author.toUpperCase();  
+            return itemData.indexOf(textData) > -1;
+        });
+        setFilterData(newFilteredData);
+    }
+
+    const bottomMenu = showSearchBar
         ? <SearchBar 
-            onPressCancel = {()=>setShowSearchBar(false)}
-            // onChangeHandler = {(text)=>setFilterText(text)}
+            onPressCancel = {()=>{setShowSearchBar(false); filterByAuthor('');}}
             onChangeHandler = {(text)=>{setFilterText(text); filterByAuthor(text);}}
             value={filterText}
         />
@@ -58,48 +65,7 @@ export default ListCompHooks = (props) =>{
         onPressSearch={()=>setShowSearchBar(true)}
         onPressSortAuthor={sortDataByAuthor}
         onPressSortId={sortDataById}
-        />);
-        setFilterText('');
-        filterByAuthor('');
-    }, [showSearchBar]);
-
-    // useEffect(()=>{
-    //     if(data!==null) filterByAuthor(filterText);
-    //     // console.warn(filterText);
-    // }, [filterText]);
-
-    //sorting
-
-    const sortDataById=()=>{
-        const newData = [...data];
-        newData.sort((a,b) => a.id - b.id);
-
-        setData(newData);
-        setFilterData(data);
-    }
-
-    const sortDataByAuthor=()=>{
-        const newData = [...data];
-        newData.sort((a,b) => {
-            if(a.author > b.author) return 1;
-            else if(a.author < b.author) return -1;
-            else return 0;
-        });
-
-        setData(newData);
-        setFilterData(newData);
-    }    
-
-    //filtering
-    const filterByAuthor=(text)=>{
-        const newData = [...data];
-        const newFilteredData = newData.filter(item =>{
-            const textData = text.toUpperCase();
-            const itemData = item.author.toUpperCase();    
-            return itemData.indexOf(textData) > -1;
-        });
-        setFilterData(newFilteredData);
-    }
+        />;
 
     return(
         <View style={styles.container}>
@@ -109,7 +75,7 @@ export default ListCompHooks = (props) =>{
                 contentContainerStyle = {styles.list}
                 renderItem={({item}) => <ListItem id={item.id} imageUrl={item.download_url} name={item.author} pageUrl={item.url}/>}
                 refreshing={isLoading}
-                onRefresh={()=>setDataRequest(!dataRequest)}
+                onRefresh={props.fetchDataMethod}
                 ListEmptyComponent = {<EmptyListComponent/>}
             />
             </View>
